reset back to events
